+++ django_echarts/entities/base.py
@@ -1,5 +1,3 @@
-
-
 class DwString(str):
     """This is a str for django template string."""