+++ tests/test_widgets.py
@@ -1,13 +1,13 @@
 import unittest
 
-from django_echarts.entities import Nav, Menu, LinkItem
+from django_echarts.entities import Nav, LinkItem
 
 
 class NavTestCase(unittest.TestCase):
     def test_nav(self):
         nav = Nav()
-        nav.add_menu('Menu1', slug='menu1', url='/menu1')
-        nav.add_item('Menu1', LinkItem('sub1', '/sub/', after_separator=True))
+        nav.add_left_menu('Menu1', slug='menu1', url='/menu1')
+        nav.add_item_in_left_menu('Menu1', LinkItem('sub1', '/sub/', after_separator=True))
 
         self.assertEqual(1, len(nav.menus))
         self.assertEqual('sub1', nav.menus[0].children[0].text)