+++ docs/changelog.md
@@ -1,5 +1,18 @@
 # 更新日志
 
+## v0.6.0
+
+> pycharts更新至2.0.x，echarts更新至5.4.x。新增 Django 4.1 构建支持。
+
+- 新增参数化图表支持
+- 新增资源类 `EntityURI`
+- 新增 `WidgetNotRegisteredError` 异常
+- `WidgetGetterMixin` 接口类支持 `EntityURI`
+- 修改导航栏(Nav)配置方式
+- 右侧导航栏支持二级菜单
+- Borax 更新至 4.0
+- typing_extensions 更新至 4.5
+
 ## v0.5.2 - （20220508）
 
 - 新增 `ElementEntity` HTML组件